Refactor the backend structure
